Added a "Scale Radius Factor" applied during meshing.
One of the problems with using Blender's meta ball constructors is that
the radius is difficult to control. The radius given shows up as the
orange outer ring, but the actual meta surface is often inside that
ring.

This commit adds a "Scale Radius Factor" which is applied to the radius
to compensate for the variability of the meta surface. It defaults to 1,
but in practice, a value of 1.4 or 1.5 seems to correct the meta surface
to better match the requested radius.