Added a custom doxygen header to link back to the main page.
The Doxyfile configuration file can include an HTML_HEADER field to
include a custom HTML header to be used when doxygen generates HTML.

To obtain a template for the header (and footer and stylesheet) use:

   doxygen -w html header.html footer.html stylesheet.css Doxyfile

That will generate what Doxygen would have used for its header, footer,
and stylesheet code in those files. The header.html file was then copied
to custom_header.html and edited to link the logo to the main page. The
main page is currently hard-coded to the github site. Note that since
Doxygen seems to generate all of its files at the top level, the use of
a relative "../../.." reference might make this more portable.